disco mode on GPU test script; send depth frames on Coppa

+++ pysrc/Coppa.py
@@ -8,8 +8,8 @@ def encodeColorFrame(colorFrame):
 	jpgTxt = base64.b64encode(buf)
 	return jpgTxt
 
-def decodeColorFrame(colorTxt):
+def decodeColorFrame(colorTxt, dtype=np.uint8):
 	imgData = base64.b64decode(colorTxt)
-	imgData = np.frombuffer(imgData, dtype=np.uint8)
+	imgData = np.frombuffer(imgData, dtype=dtype)
 	img = cv.imdecode(imgData, flags=1)
 	return img
